+++ NEWS.md
@@ -1,8 +1,9 @@
-## explore 0.6.3 (2021-01-15)
+## explore 0.7.0 (2021-01-20)
 
 * add weights parameter to explain_tree()
 * flip no/yes label in rpart.plot
 * fix minsplit for count-data
+* add weight_target()
 
 ## explore 0.6.2 (2020-10-13)